pyproject.toml
[tool.poetry]
name = "cursor-study"
version = "0.1.0"
authors = ["Hao He <haohe@andrew.cmu.edu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.0"
pandas = "^2.1.0"
matplotlib = "^3.8.0"
jupyter = "^1.1.1"
jupyterlab = "^4.3.6"
PyGithub = "^2.1.1"
python-dotenv = "^1.0.0"
gitpython = "^3.1.44"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
